// packages/project_fus/src/pages/detail/CardContent/BaseInfo/data.d.ts
/*
 * @Descripttion: 基本信息数据格式
 * @Author: linkenzone
 * @Date: 2020-09-15 14:35:04
 */

export interface PatientDataType {
  account: string;
  birthday: string;
  finishFollowup: string;
  gender: string;
  hospitalNumber: string;
  id: number;
  idNumber: string;
  nextFollowupTime: string;
  patNumber: string;
  patientName: string;
  phoneNumber1: string;
  phoneNumber2: string;
  researchCenter: string;
  updateTime: string;
}